+++ lib/maru/utils.ex
@@ -2,6 +2,11 @@ defmodule Maru.Utils do
   @moduledoc false
 
   @doc false
+  def is_blank(s) do
+    s in [nil, "", '', %{}]
+  end
+
+  @doc false
   def upper_camel_case(s) do
     s |> String.split("_") |> Enum.map(
       fn i -> i |> String.capitalize end