+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 [Upgrade Instructions](https://maru.readme.io/docs/upgrade-instructions-from-v09) From v0.9 to v0.10
 
 ## v0.10.5-dev
+* Enhancements
+  * add :keep_blank option for params
 * Bugfix
-  * allow set optional params to nil and false by default
+  * allow set optional params to false and blank value by default
 
 ## v0.10.4 (2016-8-19)
 * Enhancements